# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OnlineClipboard API
  version: 0.1.0-design
  description: v1 HTTP 契约。健康检查、就绪检查、server-info 及账号/保险库/密文/同步/设备接口已实现；客户端加密，服务端只存密文。
servers:
- url: https://clip.example.com
  description: 替换为用户的服务器 HTTPS origin
security:
- bearerAuth: []
paths:
...
  /api/v1/server-info:
    get:
      operationId: serverInfo
      summary: 服务能力发现
      x-implementation: implemented
      responses:
        '200':
          description: 服务能力发现
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerInfo'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /api/v1/auth/register:
    post:
      operationId: register
      summary: 邀请注册与初始设备会话
      x-implementation: implemented
      responses:
        '201':
          description: 邀请注册与初始设备会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/v1/auth/email-code:
    post:
      operationId: requestEmailCode
      summary: 发送邮箱注册验证码
      x-implementation: implemented
      responses:
        '204':
          description: 验证码已发送
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailCodeRequest'
  /api/v1/auth/login:
    post:
      operationId: login
      summary: 密码登录
      x-implementation: implemented
      responses:
        '200':
          description: 密码登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: 刷新并轮换令牌
      x-implementation: implemented
      responses:
        '200':
          description: 刷新并轮换令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: 撤销当前会话
      x-implementation: implemented
      responses:
        '204':
          description: 撤销当前会话
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/password:
    post:
      operationId: changePassword
      summary: 修改密码并撤销全部会话
      x-implementation: implemented
      responses:
        '204':
          description: 修改密码并撤销全部会话
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
  /api/v1/devices:
    get:
      operationId: listDevices
      summary: 列出账号设备
      x-implementation: implemented
      responses:
        '200':
          description: 列出账号设备
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevicePage'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/devices/{id}:
    delete:
      operationId: revokeDevice
      summary: 撤销设备
      x-implementation: implemented
      responses:
        '204':
          description: 撤销设备
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - $ref: '#/components/parameters/DeviceId'
  /api/v1/vault:
    get:
      operationId: getVault
      summary: 获取账号保险库信封
      x-implementation: implemented
      responses:
        '200':
          description: 获取账号保险库信封
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultEnvelope'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: initializeVault
      summary: 仅首次初始化保险库
      x-implementation: implemented
      responses:
        '201':
          description: 仅首次初始化保险库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VaultEnvelope'
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '501':
          description: 当前骨架尚未实现此业务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '428':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultEnvelope'
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
  /api/v1/vault/password-wrap:
    put:
      operationId: putPasswordWrap
      summary: 上传或替换密码包装信封
      x-implementation: implemented
      responses:
        '204':
          description: 已保存密码包装信封
        default:
          description: 统一错误体；参见 docs/06-api.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 具体业务错误码见接口设计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordWrap'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 不透明随机访问令牌，不是 JWT
...
  schemas:
...
    ServerInfo:
      type: object
      additionalProperties: false
      required:
      - name
      - version
      - stage
      - protocol_version
      - sync_available
      - e2ee_available
      - registration_mode
      - email_verification
      - min_password_chars
      - password_wrap
      - max_text_bytes
      - max_request_bytes
      - trash_retention_seconds
      properties:
        name:
          type: string
        version:
          type: string
        stage:
          type: string
          enum:
          - skeleton
          - ready
        protocol_version:
          type: integer
          enum:
          - 1
        sync_available:
          type: boolean
        e2ee_available:
          type: boolean
        registration_mode:
          type: string
          enum:
          - disabled
          - invite
          - open
          - email
          - external
        email_verification:
          type: boolean
        min_password_chars:
          type: integer
        password_wrap:
          type: boolean
        external_register_url:
          type: string
        max_text_bytes:
          type: integer
          example: 65536
        max_request_bytes:
          type: integer
          example: 131072
        trash_retention_seconds:
          type: integer
          enum:
          - 604800
...
    EmailCodeRequest:
      type: object
      additionalProperties: false
      required:
      - email
      properties:
        email:
          type: string
          format: email
...